chore(e2e): enable e2e tests on merge

See merge request Fiji/Fiji!318

+++ autoDevOps/5.E2E.sh
@@ -1,13 +1,27 @@
 #!/usr/local/bin/bash
-echo '====Start E2E'
 
+if [ "MERGE" != "${gitlabActionType}" ]
+then
+  echo "Skip e2e for action : ${gitlabActionType}"
+  exit 0
+fi
+
+echo '====Start E2E===='
 cd $project/tests/e2e/testcafe
 git clean -xdf
 npm i
 
 export SITE_URL=$appUrl
 export SCREENSHOTS_PATH=./screenshots
-# SELENIUM_SERVER is setup in jenkins as we may want to change it any time without update the code
+export SELENIUM_SERVER=$SELENIUM_SERVER
+export BROWSERS=$BROWSERS
+export RC_PLATFORM_APP_KEY=$RC_PLATFORM_APP_KEY
+export RC_PLATFORM_APP_SECRET=$RC_PLATFORM_APP_SECRET
+export ACTION="ON_${gitlabActionType}"
+export BRANCH="${gitlabBranch}"
+export ENABLE_REMOTE_DASHBOARD=true
+
+echo "Following environment variables are used by e2e tests to define test scope:"
 echo $SELENIUM_SERVER
 echo $BROWSERS
 echo $RC_PLATFORM_APP_KEY
@@ -18,20 +32,11 @@ echo $BRANCH
 echo $SCREENSHOTS_PATH
 
 mkdir -p $SCREENSHOTS_PATH
+npx ts-node create-run-id.ts
 npx ts-node multi-run.ts
-rm -r $SCREENSHOTS_PATH
-
 e2eResult=$?
 
-xmlResult=$project/tests/e2e/testcafe/allure/allure-results
-htmlResult=$project/tests/e2e/testcafe/allure/html-results
-allure generate $xmlResult -o $htmlResult
-
-localFolder=$htmlResult
-remoteFolder=e2e/$subDomain/$BUILD_NUMBER
-syncFolderToServer $localFolder $remoteFolder
-
-E2EUrl=https://e2e.fiji.gliprc.com/$subDomain/$BUILD_NUMBER/index.html
-addEnv E2EResult="> **E2E Url**: $E2EUrl"
+export REPORT_URL=`cat ./reportUrl`
+addEnv E2EResult="> **E2E Url**: $REPORT_URL"
 
 echo "E2E Result: $e2eResult"
